+++ themes/besocial/css/_besocial/icons_svg_besocial.css
@@ -16,6 +16,10 @@
 	--hexcode: #FFA500;
 }
 
+.i-last_post::before {
+	content: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' fill='%23000000' viewBox='0 0 38 38'%3E%3Cpath d='M.8 29.5c-.4.4-.8.3-.8-.4V3c0-.8.4-1 .8-.5l12.7 12.7.2.3V3c0-.8.4-1 .8-.5l12.7 12.7.2.3v-12c0-.7.5-1.2 1.2-1.2H31c.5 0 1 .5 1 1v25.3c0 .6-.5 1-1 1h-2.4c-.7 0-1.2-.4-1.2-1v-12l-.2.2-12.7 12.7c-.4.4-.8.3-.8-.4V16.6s0 .2-.2.3z'/%3E%3C/svg%3E");
+}
+
 .i-menu-pm-off::before,
 .i-menu-mentions-off::before,
 .i-menu-admin::before,
